discord-bot/requirements.txt
discord.py==2.3.2
mysql-connector-python==8.2.0
aiohttp==3.9.1
pytz==2023.3
